Move plugin updater
