+++ plugin/addon_updater_ops.py
@@ -31,7 +31,7 @@ from bpy.app.handlers import persistent
 # Prevents popups for users with invalid python installs e.g. missing libraries
 # and will replace with a fake class instead if it fails (so UI draws work).
 try:
-    from .addon_updater import Updater as updater
+    from plugin.addon_updater import Updater as updater
 except Exception as e:
     print("ERROR INITIALIZING UPDATER")
     print(str(e))
